pyproject.toml
[project]
authors = [
  { name = "Konstantin Borovik", email = "kb@lab5.ca" },
]
dependencies = [
  "google-genai>=0.5.0",
  "logfire>=3.2.0",
  "pdf2image>=1.17.0",
  "python-magic>=0.4.27",
  "xero-python>=6.3.0",
]
description = "Convert unstructured data into structured data for Xero Accounting"
name = "xero-ai"
readme = "README.md"
requires-python = ">=3.12"
version = "2025.01.26.post1946"

[project.scripts]
xero-ai = "xero_ai:main"

[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "python-dotenv>=1.0.1",
  "ruff>=0.9.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [ "C90", "F", "N", "N", "PL", "RUF", "SIM", "UP" ]

[tool.ruff.lint.pydocstyle]
convention = "google"
